tests/submission/PhilipTaylor/canvas/2d.fillStyle.get.transparent.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.fillStyle.get.transparent</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.fillStyle.html">fillStyle</a>.<a href="index.2d.fillStyle.get.html">get</a>.transparent</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.colours.getcolour">2d.colours.getcolour</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.serializecolour.transparent">2d.serializecolour.transparent</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.fillStyle = 'rgba(0,0,0,0)';
_assertEqual(ctx.fillStyle, 'rgba(0, 0, 0, 0)', "ctx.fillStyle", "'rgba(0, 0, 0, 0)'");


});
</script>

<div id="log"></div>
</body></html>
